Update jupiter dependency to support async
This change was not as simple as updating a version number,
unfortunately. To get access to async/await, you have to grab jupiter
directly from GitHub from either the async/await or master branch
because the newest official release is 0.14.2. On the master branch,
aside from async/await features, jupiter now uses version 0.8 of the
uuid crate, which is great! That means we must update our own uuid
dependency to v0.8 to make our GraphQL types that use Uuid compatible
with jupiter... but for some reason, this completely breaks sqlx-macros
when the uuid feature is enabled for sqlx - when compiling the project,
sqlx-macros will spew out linking errors.

So jupiter needs v0.8 of uuid, and sqlx will only compile on v0.7 of
uuid. How do we fix this? The current solution is to make a fork of
jupiter (KongBorup/jupiter) where uuid has been downgraded to v0.7,
and now both jupiter and sqlx compile just fine.

Besides problems with sqlx, upgrading jupiter also led to some
dependency conflicts with actix-web (regarding async-traits), which has
been fixed by upgrading actix-web to version 3, which was a due change
anyway.

One more integration test has also been added to check the apiVersion
query.