Clean up dependencies
This is done to rectify the changes introduced when trying to fix the
sqlx-macros linking errors.

This commit also removes unused rusqlite dependencies, but there may be
more remnants still around. They will be removed later.

As of this commit, the project should work properly with uuid v0.8,
juniper's master branch, actix-web v3, and sqlx v0.4.